Todos can be marked as checked.

+++ app/Controllers/TodosController.js
@@ -1,7 +1,7 @@
 import { appState } from "../AppState.js";
 import { todosService } from "../Services/TodosService.js";
 import { Pop } from "../Utils/Pop.js";
-import { setHTML } from "../Utils/Writer.js";
+import { setHTML, setText } from "../Utils/Writer.js";
 
 function _drawTodos() {
   let todos = appState.todos
@@ -10,16 +10,25 @@ function _drawTodos() {
   setHTML('todo-list', template)
 }
 
+function _drawCount() {
+  let count = 0
+  appState.todos.forEach(t => t.completed ? count++ : '')
+  setText('todo-count', count)
+}
+
 export class TodosController {
   constructor() {
     this.getTodos()
     appState.on('todos', _drawTodos)
+    appState.on('todos', _drawCount)
+
   }
 
 
   async getTodos() {
     try {
       await todosService.getTodos()
+      _drawCount()
     } catch (error) {
       console.error(error)
       Pop.error(error)
@@ -31,6 +40,8 @@ export class TodosController {
       let form = window.event.target
       // @ts-ignore
       let description = form.description.value
+      // @ts-ignore
+      form.reset()
       await todosService.addTodo(description)
     } catch (error) {
       console.error(error)
@@ -41,7 +52,19 @@ export class TodosController {
   }
   async delete(id) {
     try {
-      await todosService.delete(id)
+      if (await Pop.confirm('Are you sure you want to remove this todo?')) {
+        await todosService.delete(id)
+        Pop.toast('todo removed!', 'success', 'top', 1500)
+      }
+    } catch (error) {
+      console.error(error)
+      Pop.error(error)
+    }
+  }
+
+  async complete(id) {
+    try {
+      await todosService.complete(id)
     } catch (error) {
       console.error(error)
       Pop.error(error)
